+++ R/ViewShiny.R
@@ -127,9 +127,9 @@ prepareCharacterizationShiny <- function(
 
 viewChars <- function(
     databaseSettings,
-    testApp = F) {
-  ensure_installed("ShinyAppBuilder")
-  ensure_installed("ResultModelManager")
+    testApp = F
+    ) {
+  ensure_installed("OhdsiShinyAppBuilder")
 
   connectionDetails <- do.call(
     DatabaseConnector::createConnectionDetails,
@@ -138,29 +138,7 @@ viewChars <- function(
   connection <- ResultModelManager::ConnectionHandler$new(connectionDetails)
   databaseSettings$connectionDetailsSettings <- NULL
 
-  if (utils::packageVersion("ShinyAppBuilder") < "1.2.0") {
-    # use old method
-    # set database settings into system variables
-    Sys.setenv("resultDatabaseDetails_characterization" = as.character(ParallelLogger::convertSettingsToJson(databaseSettings)))
-
-    config <- ParallelLogger::loadSettingsFromJson(
-      fileName = system.file(
-        "shinyConfig.json",
-        package = "Characterization"
-      )
-    )
-
-    if (!testApp) {
-      ShinyAppBuilder::viewShiny(
-        config = config,
-        connection = connection
-      )
-    } else {
-      ShinyAppBuilder::createShinyApp(config = config, connection = connection)
-    }
-  } else {
     # use new method
-
     config <- ParallelLogger::loadSettingsFromJson(
       fileName = system.file(
         "shinyConfigUpdate.json",
@@ -171,23 +149,13 @@ viewChars <- function(
     databaseSettings$cgTablePrefix <- databaseSettings$cohortTablePrefix
     databaseSettings$databaseTable <- "DATABASE_META_DATA"
     databaseSettings$databaseTablePrefix <- ""
-    # databaseSettings$iTablePrefix <- databaseSettings$incidenceTablePrefix
     databaseSettings$cgTable <- "cohort_definition"
 
-    if (!testApp) {
-      ShinyAppBuilder::viewShiny(
+      OhdsiShinyAppBuilder::viewShiny(
         config = config,
         connection = connection,
         resultDatabaseSettings = databaseSettings
       )
-    } else {
-      ShinyAppBuilder::createShinyApp(
-        config = config,
-        connection = connection,
-        resultDatabaseSettings = databaseSettings
-      )
-    }
-  }
 }
 
 
@@ -223,19 +191,7 @@ ensure_installed <- function(pkg) {
     if (interactive()) {
       message(msg, "\nWould you like to install it?")
       if (utils::menu(c("Yes", "No")) == 1) {
-        if (pkg %in% c("ShinyAppBuilder", "ResultModelManager")) {
-          # add code to check for devtools...
-          dvtCheck <- tryCatch(utils::packageVersion("devtools"),
-            error = function(e) NA
-          )
-          if (is.na(dvtCheck)) {
-            utils::install.packages("devtools")
-          }
-
-          devtools::install_github(paste0("OHDSI/", pkg))
-        } else {
           utils::install.packages(pkg)
-        }
       } else {
         stop(msg, call. = FALSE)
       }